# api-documentation.yaml
...
openapi: "3.0.0"
info:
  title: "Freelancer Backend API"
  description: "**Untitled-repo**\n\nThis is the OpenAPI documentation of our backend\
    \ API.\n"
  version: "1.0.0"
servers:
  - url: "https://untitled-repo-backend.herokuapp.com/api/v1"
    description: "The URL of the server"
  - url: "https://virtserver.swaggerhub.com/untitled-repo/FreelancerBackendAPI/1.0.0"
    description: "SwaggerHub API Auto Mocking"
...
paths:
...
  /jobs:
    get:
      tags:
        - "jobs"
      summary: "Lists available Jobs"
      description: "Lists the available jobs"
      operationId: "getAllAvailableJobs"
      parameters:
        - name: "size"
          in: "query"
          description: "The filter on the size of the packages"
          required: false
          style: "form"
          explode: true
          schema:
            type: "string"
            enum:
              - "small"
              - "medium"
              - "large"
      responses:
        "200":
          description: "The matching job array"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Job"
            application/xml:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Job"
        "400":
          description: "Invalid format"
        "401":
          description: "Access token is missing or invalid"
        "404":
          description: "No Jobs available"
      security:
        - bearerAuth: []
    post:
      tags:
        - "jobs"
      summary: "Posts a Job"
      operationId: "postJob"
      parameters:
        - name: "user-id"
          in: "query"
          description: "ID of the User that post's the Job"
          required: true
          style: "form"
          explode: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        $ref: "#/components/requestBodies/NewJob"
      responses:
        "201":
          description: "Created"
          headers:
            Location:
              style: "simple"
              explode: false
              schema:
                type: "string"
                format: "uri"
                example: "jobs/post/2574120"
        "405":
          description: "Invalid input"
      security:
        - bearerAuth: []
  /jobs/post/{job-id}:
    get:
      tags:
        - "jobs"
      summary: "Get Job from Job ID"
      description: "Returns a single Job"
      operationId: "getJobById"
      parameters:
        - name: "job-id"
          in: "path"
          description: "ID of the Job"
          required: true
          style: "simple"
          explode: false
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "Information on the queried Job"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
            application/xml:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "Access token is missing or invalid"
        "404":
          description: "Job not found"
      security:
        - bearerAuth: []
    put:
      tags:
        - "jobs"
      summary: "Update a Job"
      description: "Only the User who posted the Job can update it"
      operationId: "updateJob"
      parameters:
        - name: "job-id"
          in: "path"
          description: "ID of the Job that needs to be updated"
          required: true
          style: "simple"
          explode: false
          schema:
            type: "integer"
            format: "int64"
        - name: "user-id"
          in: "query"
          description: "ID of the User who posted the Job"
          required: true
          explode: false
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        $ref: "#/components/requestBodies/NewJob"
      responses:
        "200":
          description: "OK"
        "401":
          description: "Access token is missing or invalid"
        "404":
          description: "Job not found"
        "405":
          description: "Invalid input"
        "409":
          description: "You couldn't modify the Job"
      security:
        - bearerAuth: []
    delete:
      tags:
        - "jobs"
      summary: "Deletes a Job"
      description: "Only the User who posted the Job can delete it"
      operationId: "deleteJob"
      parameters:
        - name: "job-id"
          in: "path"
          description: "Job ID to delete"
          required: true
          style: "simple"
          explode: false
          schema:
            type: "integer"
            format: "int64"
      responses:
        "400":
          description: "Invalid id supplied"
        "401":
          description: "Access token is missing or invalid"
        "404":
          description: "Job not found"
        "409":
          description: "You couldn't delete the Job"
      security:
        - bearerAuth: []
  /jobs/accept/{job-id}:
    post:
      tags:
        - "jobs"
      summary: "Accept a Job"
      description: "The User who posted it can't accept their own Jobs."
      operationId: "acceptJob"
      parameters:
        - name: "job-id"
          in: "path"
          description: "Job ID to delete"
          required: true
          style: "simple"
          explode: false
          schema:
            type: "integer"
            format: "int64"
        - name: "user-id"
          in: "query"
          description: "The deliverer user's id"
          required: true
          style: "form"
          explode: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "400":
          description: "Invalid id supplied"
        "401":
          description: "Access token is missing or invalid"
        "404":
          description: "Job not found"
        "409":
          description: "Couldn't accept Job"
      security:
        - bearerAuth: []
    delete:
      tags:
        - "jobs"
      summary: "Abandon a Job"
      description: "The User who currently accepted the Job can abandon it."
      operationId: "abandonJob"
      parameters:
        - name: "job-id"
          in: "path"
          description: "Job ID to delete"
          required: true
          style: "simple"
          explode: false
          schema:
            type: "integer"
            format: "int64"
        - name: "user-id"
          in: "query"
          description: "The deliverer user's id"
          required: true
          style: "form"
          explode: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "400":
          description: "Invalid id supplied"
        "401":
          description: "Access token is missing or invalid"
        "404":
          description: "Job not found"
      security:
        - bearerAuth: []
components:
  schemas:
...
    Job:
      required:
        - "deadline"
        - "sender-id"
        - "deliveryRoute"
        - "job-id"
        - "jobIssuedDate"
        - "payment"
        - "size"
        - "status"
      type: "object"
      properties:
        job-id:
          type: "integer"
          description: "The ID of the Job"
          format: "int64"
          nullable: false
        sender-id:
          type: "integer"
          description: "The ID of the sender User"
          format: "int64"
          nullable: false
        status:
          type: "string"
          description: "The current status of the Job"
          nullable: false
          enum:
            - "pending"
            - "accepted"
            - "pickedUp"
            - "delivered"
            - "expired"
        senderRating:
          maximum: 5
          minimum: 1
          type: "integer"
          description: "The sender rating of the delivery"
          nullable: true
        size:
          type: "string"
          description: "The size of the package"
          nullable: false
          enum:
            - "small"
            - "medium"
            - "large"
        payment:
          type: "integer"
          description: "The payment for the delivery"
          nullable: false
        jobIssuedDate:
          type: "string"
          description: "The date the sender issued the Job"
          format: "date-time"
          nullable: false
        deadline:
          type: "string"
          description: "The deadline of the Job"
          format: "date-time"
          nullable: false
        deliveryDate:
          type: "string"
          description: "The day the package was delivered"
          format: "date-time"
          nullable: true
        deliveryRoute:
          $ref: "#/components/schemas/Route"
      description: "Detailed information on a Job"
      xml:
        name: "Job"
    JobRegistration:
      required:
        - "deadline"
        - "destination"
        - "jobIssuedDate"
        - "payment"
        - "size"
        - "startLocation"
      type: "object"
      properties:
        size:
          type: "string"
          description: "The size of the package"
          nullable: false
          enum:
            - "small"
            - "medium"
            - "large"
        payment:
          type: "integer"
          description: "The payment for the delivery"
          nullable: false
        jobIssuedDate:
          type: "string"
          description: "The date the sender issued the Job"
          format: "date-time"
          nullable: false
        deadline:
          type: "string"
          description: "The deadline of the Job"
          format: "date-time"
          nullable: false
        startLocation:
          type: "string"
          description: "The start location of the Route"
          nullable: false
        destination:
          type: "string"
          description: "The destination of the Route"
          nullable: false
      description: "Data on the new User"
      xml:
        name: "JobRegistration"
    Route:
      required:
        - "destination"
        - "startLocation"
      type: "object"
      properties:
        actualTime:
          type: "integer"
          description: "The time the Route was completed"
          format: "int64"
          nullable: true
        optimalTime:
          type: "integer"
          description: "The optimal time for the Route"
          format: "int64"
          nullable: true
        startLocation:
          type: "string"
          description: "The start location of the Route"
          nullable: false
        destination:
          type: "string"
          description: "The destination of the Route"
          nullable: false
      description: "A Route for a specific Job"
      xml:
        name: "Route"
...
  requestBodies:
...
    NewJob:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JobRegistration"
        application/xml:
          schema:
            $ref: "#/components/schemas/JobRegistration"
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"